initial homepage layout complete
