+++ src/content/homepage/index.md
@@ -1,6 +1,6 @@
 ---
-#hero
-hero:
+#HOME INTRO
+homeIntro:
   subtitle: Hello
   title: Šimon Podhajský is the resident data
   rolling_texts:
@@ -14,52 +14,62 @@ hero:
     label: Contact me
     href: "/contact/"
 
-banner:
-  title: "Šimon Podhajský is the resident data everything"
-  image: /images/bannerimage.webp
-  meta: "this is meta"
-  link:
-    label: Contact me
-    href: "/contact/"
-
-# name_image
-name_image:
-  title: My name is Czech.
-  description: The “š” is pronounced like the “sh” in “shoe”, and the “ý” is pronounced like the “ee” in “bee”. So, it’s “SHIH-mon Pot-HIGH-skee”. But you can always call me Simon.
-  image: "/images/videopopup.webp"
-
-## capabilities
-capabilities:
+## THINGS I DO
+thingsIDo:
   subtitle: SPECIAL SKILLS
   title: Things I Can Do
   description: My LinkedIn bio used to say I'm a "Data Everything". That wasn't too far off.
   list:
     - icon: "lock"
-      title: Evaluate & productionize LLM-based applications
+      title: Evaluate & Productionize LLM-based Applications
       content: I understand that integrating standard software engineering and ML best practices is what keeps the modern LLM-based apps from remaining merely a fancy demo.
     - icon: "wind"
       title: Data Engineering
       content: "Performant, semantically rich, reproducible: my pipelines are all of the above."
     - icon: "shield"
       title: Data Science and Data Analytics
-      content: (It's more than just averages and sums across the right groups.)
+      content: It's more than just averages and sums across the right groups.
     - icon: "link"
       title: Full stack, all the way to deployment.
       content: Does the project need a frontend to interact with and a cloud infrastructure to host itself in? I'm on it.
     - icon: "lock"
       title: Communication and Public Speaking.
       content: >
-        [I've spoken at local meetups, online events, and big conferences.](/presentations) **Hit me up to speak at any of the above!**
+        I've spoken at local meetups, online events, and big conferences. Hit me up to speak at any of the above!
     - icon: "wind"
       title: I'm a Generalist.
       content: ...which just means that I don't stop at a problem's boundary; I see it through wherever it
 
-# home_blog
-home_blog:
+# SIDE PROJECTS
+sideProjects:
+  subtitle: Side Projects
+  title: Sometimes, I... <br> ...work on side projects
+  link:
+    label: Explore Side Projects
+    href: "/side_projects/"
+
+# PRESENTATIONS
+presentations:
+  subtitle: Presentations
+  title: Sometimes, I... host talks and presentations
+  description: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation.
+  image: /images/about/01.jpg
+  link:
+    label: Explore Presentations
+    href: "/presentations/"
+
+# NAME IMAGE
+nameImage:
+  title: My name is Czech.
+  description: The “š” is pronounced like the “sh” in “shoe”, and the “ý” is pronounced like the “ee” in “bee”. So, it’s “SHIH-mon Pot-HIGH-skee”. But you can always call me Simon.
+  image: "/images/videopopup.webp"
+
+# ARTICLES
+articles:
   subtitle: Articles
   title: Sometimes, I... <br> ...write technical articles
   link:
-    label: View More
+    label: View More Articles
     href: "/blog/"
 
 #features
@@ -87,6 +97,14 @@ features:
     - icon: "wind"
       title: I'm a generalist.
       content: ...which just means that I don't stop at a problem's boundary; I see it through wherever it needs to go.
+
+banner:
+  title: "Šimon Podhajský is the resident data everything"
+  image: /images/bannerimage.webp
+  meta: "this is meta"
+  link:
+    label: Contact me
+    href: "/contact/"
 
 ##brands - deleted from index.astro
 brands:
