--- src/layouts/components/home/Capabilities.astro
@@ -1,35 +1,0 @@
----
-import { markdownify } from "@/lib/utils/textConverter";
-const {
-  data: { subtitle, title, list },
-} = Astro.props;
----
-
-<section class="section">
-  <div class="container">
-    <div class="row justify-center">
-      <div class="col-12">
-        <div class="section-title">
-          <p
-            set:html={markdownify(subtitle)}
-            class="mb-4 text-[0.9rem] uppercase"
-          />
-          <h2 set:html={markdownify(title)} class="title" />
-        </div>
-      </div>
-      {
-        list.map((item: any) => (
-          <div class="col-12 mb-4 md:col-6 lg:col-4 p-4">
-            <div class="mt-6 px-0 text-center ">
-              <h3 class="mb-3 text-h5  font-semibold">{item.title}</h3>
-              <p
-                class="text-sm leading-6"
-                set:html={markdownify(item.content)}
-              />
-            </div>
-          </div>
-        ))
-      }
-    </div>
-  </div>
-</section>
