+++ src/layouts/partials/Cta.astro
@@ -23,7 +23,7 @@ const {
                   <h2 set:html={markdownify(title)} class="title" />
                   <p
                     set:html={markdownify(content)}
-                    class="text-[.9rem] text-text"
+                    class="text-base 2xl:text-lg text-text"
                   />
                 </div>
