+++ src/layouts/partials/Header.astro
@@ -19,8 +19,7 @@ const { main }: { main: NavigationLink[] } = menu;
 <header
   class="header"
   id="header"
-  data-aos="fade-down"
-  data-aos-anchor-placement="top-bottom"
+  data-aos="fade-up"
   data-aos-easing="ease-in-out"
   data-aos-duration="1000"
 >
