+++ src/styles/components.scss
@@ -125,7 +125,7 @@
 .section-title {
   @apply mb-10 text-center;
   .title {
-    @apply relative mb-10 text-h2-sm font-semibold capitalize leading-[1.3];
+    @apply relative mb-10 text-h2-sm font-bold capitalize leading-[1.3];
     &::after {
       @apply absolute -bottom-5 left-[50%] h-[5px] w-[4.3rem] -translate-x-[50%] rounded-[1.8rem] bg-primary content-[""];
     }
@@ -139,7 +139,7 @@
 }
 
 .features-card {
-  @apply mx-[0.9rem] my-5 rounded-md border border-solid border-transparent px-[1.875rem] py-[4.3rem] text-center shadow-[0px_4px_25px_rgba(0,0,0,.05)] duration-300 ease-[ease] hover:-translate-y-[2px] hover:border-primary/10 hover:shadow-none;
+  @apply mx-[0.9rem] my-5 rounded-md border border-solid border-transparent px-[1.875rem] py-[4.3rem] text-center shadow duration-300 ease-[ease] hover:-translate-y-[2px] hover:border-primary/10 hover:shadow-none h-[400px];
 }
 .icon {
   @apply relative z-10 m-auto block h-20 w-20 overflow-hidden rounded-[0.9rem] border border-solid border-primary/10 text-h2 leading-[5rem] duration-150;
